# IdleSchemes.Api/openapispec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IdleSchemes.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/events/all/{eventId}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "GetEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventInstanceModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventInstanceModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventInstanceModel"
                }
              }
            }
          }
        }
      }
    },
    "/events/all/{eventId}/tickets": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "GetAvailableTickets",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketClassModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketClassModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketClassModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/events/regions/{regionId}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "GetEventsInRegion",
        "parameters": [
          {
            "name": "regionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": null
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventInstanceModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventInstanceModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventInstanceModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/events/orgs/{organizationId}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "GetEventsByOrg",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": null
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventInstanceModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventInstanceModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventInstanceModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/events/associates/{associateId}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "GetEventsByAssociate",
        "parameters": [
          {
            "name": "associateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": null
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventInstanceModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventInstanceModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventInstanceModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/organizations/all/{organizationId}": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "GetOranization",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationModel"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/all/{organizationId}/associates": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "GetAssociatesInOrganization",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssociateModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssociateModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssociateModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/organizations/byregion/{regionId}": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "GetOranizationsInRegion",
        "parameters": [
          {
            "name": "regionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationModel"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssociateModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "publicBio": {
            "type": "string"
          }
        }
      },
...
      "EventInstanceModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageInfoModel"
            }
          }
        }
      },
      "ImageInfoModel": {
        "type": "object",
        "properties": {
          "src": {
            "type": "string"
          },
          "caption": {
            "type": "string"
          }
        }
      },
...
      "OrganizationModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "SeatSelection": {
        "type": "integer"
      },
...
      "TicketClassAvailabilityModel": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "confirmedCount": {
            "type": "integer",
            "format": "int32"
          },
          "pendingCount": {
            "type": "integer",
            "format": "int32"
          },
          "confirmedAndPendingCount": {
            "type": "integer",
            "format": "int32"
          },
          "remainingCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "remainingSeats": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pendingSeats": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "nullable": true
      },
      "TicketClassModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "basePrice": {
            "type": "number",
            "format": "double"
          },
          "availability": {
            "$ref": "#/components/schemas/TicketClassAvailabilityModel"
          },
          "seatSelection": {
            "$ref": "#/components/schemas/SeatSelection"
          }
        }
      },
...
    }
  },
...
}